requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
boto3==1.34.0
pydantic==2.5.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
streamlit==1.28.0
plotly==5.17.0
